pyproject.toml
[tool.poetry]
name = "webook-html-generator"
version = "0.1.0"
description = ""
authors = ["zmarusic <zeljkomarusic@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
Jinja2 = "^3.0.3"
pedantic = "^1.9.0"
python-dotenv = "^0.19.2"
pydantic = {extras = ["email"], version = "^1.9.0"}
pyhumps = "^3.5.3"
APScheduler = "^3.9.1"
python-slugify = "^6.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
